+++ setup.cfg
@@ -38,7 +38,7 @@ OPTIONAL =
     # Hydra
     hydra-core>=1.3.2
     omegaconf>=2.3.0
-    # Miscaleaneous
+    # Miscellanous
     tqdm>=4.66.3
 
 # Required for testing
